extract Sliders
