+++ platform/jewel/samples/standalone/src/main/kotlin/org/jetbrains/jewel/samples/standalone/viewmodel/ComponentsViewModel.kt
@@ -22,7 +22,7 @@ import org.jetbrains.jewel.samples.showcase.components.RadioButtons
 import org.jetbrains.jewel.samples.showcase.components.Scrollbars
 import org.jetbrains.jewel.samples.showcase.components.SegmentedControls
 import org.jetbrains.jewel.samples.showcase.components.StandaloneSampleIcons
-import org.jetbrains.jewel.samples.standalone.view.component.Sliders
+import org.jetbrains.jewel.samples.showcase.components.Sliders
 import org.jetbrains.jewel.samples.standalone.view.component.SplitLayouts
 import org.jetbrains.jewel.samples.standalone.view.component.Tabs
 import org.jetbrains.jewel.samples.standalone.view.component.TextAreas
